backend/requirements.txt
fastapi
uvicorn
motor
joblib
pandas
scikit-learn
python-dotenv